feat(routes) users routes created and moved it to a new folder: routes
