+++ db/index.js
@@ -26,7 +26,7 @@ require('./models')
 
 // sync the db, creating it if necessary
 function sync(force = app.isTesting, retries = 0, maxRetries = 5) {
-  return db.sync({force})
+  return db.sync({force:true})
     .then(() => console.log(`Synced models to db ${url}`))
     .catch(fail => {
       // Don't do this auto-create nonsense in prod, or
