+++ server/api.js
@@ -6,7 +6,7 @@ const api = module.exports = require('express').Router() // eslint-disable-line 
 api
   .get('/heartbeat', (req, res) => res.send({ok: true}))
   .use('/auth', require('./auth'))
-  .use('/users', require('./users'))
+  .use('/users', require('./routes/users'))
 
 // No routes matched? 404.
 api.use((req, res) => res.status(404).end())
