+++ server/auth.filters.js
@@ -7,6 +7,13 @@ const mustBeLoggedIn = (req, res, next) => {
   next()
 }
 
+// const mustBeAdmin = (req, res, next) => {
+//   if (!req.user.isAdmin) {
+//     return res.status(401).send('You must be an admin')
+//   }
+//   next()
+// }
+
 const selfOnly = action => (req, res, next) => {
   if (req.params.id !== req.user.id) {
     return res.status(403).send(`You can only ${action} yourself.`)
